requirements.txt
pyqt5>=5.15.0
pyqtgraph>=0.13.0
pyserial>=3.5
numpy>=1.21.0
